camera e criacao dos arquivos cpp

+++ Material.h
@@ -7,16 +7,16 @@ class Material
 {
 private:
 	float ka, kd, ks;
-	unsigned float n;
+	float n;
 	Color od;
 public:
-	Material(float ka, float kd, float ks, unsigned float n, Color od);
+	Material(float ka, float kd, float ks, float n, Color od);
 	~Material();
 	
 	float get_ka() const;
 	float get_kd() const;
 	float get_ks() const;
-	unsigned float get_n() const;
+	float get_n() const;
 	Color get_od() const;
 	
 	void set_color(Color color);
